1，修改时间刷新 2，添加仓库

+++ Classes/TopBar.h
@@ -24,7 +24,7 @@ private:
 	cocos2d::ui::TextBMFont* livedayslbl;
 	cocos2d::ui::TextBMFont* timelbl;
 	cocos2d::ui::TextBMFont* templbl;
-	int pastmin;//虚拟分钟，
+	float pastmin;//虚拟分钟，
 	ProgressTimer* outinjuryBar;
 	ProgressTimer* innerinjuryBar;
 	ProgressTimer* hungerBar;
